Sync the whole state on terminate (#1)
This PR fixes a problem when changes made right before the CRDT
terminates where not synced to the neighbours. The reason is that
syncing changes requires exchaning at least two messages with the
neighbours, and that process was not completed during CRDT termination.

The fix is very simple - when CRDT terminates, we broadcast the whole state
to the neighbours. This should be fine because currently our CRDTs in production
do not have too many entries. This PR also extends the existing telemetry event
with the total CRDT size, so that we can monitor it in production and take action
if it grows too big.

Co-authored-by: Rafal Studnicki <rafal@whatnot.com>